requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ML/AI Dependencies
transformers==4.37.2
torch==2.1.2
accelerate==0.26.1

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing FastAPI endpoints